feat(LKR-5): Add the sign-in page, structure folder features
