+++ apps/admin-web/app/auth/sign-in/page.tsx
@@ -1,0 +1,5 @@
+import { SignInForm } from "@/features/auth/components/signInForm";
+
+export default function SignInPage() {
+  return <SignInForm />;
+}
