+++ apps/admin-web/features/auth/index.ts
@@ -1,0 +1,2 @@
+export { SignUpForm } from "./components/signUpForm";
+export { SignInForm } from "./components/signInForm";
