+++ apps/admin-web/public/favicon.ico
