+++ packages/ui/src/components/phoneInput.tsx
@@ -17,7 +17,7 @@ function PhoneInput({ className, ...props }: Props) {
       inputClass={cn(
         "selection:!bg-primary selection:!text-primary-foreground dark:!bg-input/30 !border-input !flex !h-9 !w-full !min-w-0 !rounded-md !border !bg-transparent !text-sm !shadow-xs !transition-[color,box-shadow] !outline-none disabled:!pointer-events-none disabled:!cursor-not-allowed disabled:!opacity-50 md:!text-sm",
         "focus-visible:!border-ring",
-        "aria-invalid:!ring-destructive/20 dark:!aria-invalid:ring-destructive/40 aria-invalid:!border-destructive",
+        "aria-invalid:!ring-destructive/20 dark:!aria-invalid:ring-destructive/40 aria-invalid:!border-destructive aria-invalid:!bg-destructive/10 dark:aria-invalid:!bg-destructive/10",
         className,
       )}
       containerClass="[&_.country]:hover:!bg-accent [&_.highlight]:!bg-accent"
